Feature/tripschedule

+++ MSGS/src/main/java/com/msgs/tripschedule/dao/TripScheduleDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TripScheduleDAO extends JpaRepository<TripSchedule, Long> {
+public interface TripScheduleDAO extends JpaRepository<TripSchedule, Integer> {
 
 }
